+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
 	".",
-	# "chekov_macros",
+	"chekov_macros",
 	"event_store",
 	"examples/*"
 ]
