+++ src/command_executor.rs
@@ -1,0 +1,7 @@
+use crate::Aggregate;
+use crate::Command;
+use crate::CommandExecutorError;
+
+pub trait CommandExecutor<T: Command + ?Sized>: Aggregate {
+    fn execute(cmd: T, state: &Self) -> Result<Vec<T::Event>, CommandExecutorError>;
+}
